pyproject.toml
[project]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = "homework-08/fastAPI"
authors = [
    {name = "Sergey Kravchenko"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi[standard] (>=0.115.13,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "bcrypt (<4.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

